+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- `Logger.Writer`: fixed a minor bug.
 
 ## [1.0.1] - 2016-08-20
 ### Changed